remove Bucket.start_large_file and deprecated Bucket.copy_file
