+++ b2sdk/bucket.py
@@ -391,19 +391,6 @@ class Bucket(metaclass=B2TraceMeta):
             prefix=prefix,
         )
 
-    def start_large_file(self, file_name, content_type=None, file_info=None):
-        """
-        Start a large file transfer.
-
-        :param str file_name: a file name
-        :param str,None content_type: the MIME type, or ``None`` to accept the default based on file extension of the B2 file name
-        :param dict,None file_info: a file info to store with the file or ``None`` to not store anything
-        """
-        validate_b2_file_name(file_name)
-        return self.api.services.large_file.start_large_file(
-            self.id_, file_name, content_type=content_type, file_info=file_info
-        )
-
     @limit_trace_arguments(skip=('data_bytes',))
     def upload_bytes(
         self,
@@ -809,44 +796,6 @@ class Bucket(metaclass=B2TraceMeta):
                 progress_listener=progress_listener,
                 encryption=destination_encryption,
             )
-
-    # FIXME: this shold be deprecated
-    def copy_file(
-        self,
-        file_id,
-        new_file_name,
-        bytes_range=None,
-        metadata_directive=None,
-        content_type=None,
-        file_info=None,
-        destination_encryption: Optional[EncryptionSetting] = None,
-        source_encryption: Optional[EncryptionSetting] = None,
-    ):
-        """
-        Creates a new file in this bucket by (server-side) copying from an existing file.
-
-        :param str file_id: file ID of existing file
-        :param str new_file_name: file name of the new file
-        :param tuple[int,int],None bytes_range: start and end offsets (**inclusive!**), default is the entire file
-        :param b2sdk.v1.MetadataDirectiveMode,None metadata_directive: default is :py:attr:`b2sdk.v1.MetadataDirectiveMode.COPY`
-        :param str,None content_type: content_type for the new file if metadata_directive is set to :py:attr:`b2sdk.v1.MetadataDirectiveMode.REPLACE`, default will copy the content_type of old file
-        :param dict,None file_info: file_info for the new file if metadata_directive is set to :py:attr:`b2sdk.v1.MetadataDirectiveMode.REPLACE`, default will copy the file_info of old file
-        :param b2sdk.v1.EncryptionSetting destination_encryption: encryption settings for the destination
-                (``None`` if unknown)
-        :param b2sdk.v1.EncryptionSetting source_encryption: encryption settings for the source
-                (``None`` if unknown)
-        """
-        return self.api.session.copy_file(
-            file_id,
-            new_file_name,
-            bytes_range,
-            metadata_directive,
-            content_type,
-            file_info,
-            self.id_,
-            destination_server_side_encryption=destination_encryption,
-            source_server_side_encryption=source_encryption,
-        )
 
     def delete_file_version(self, file_id, file_name):
         """
